+++ pipe/src/main/java/com/swpu/pipe/entity/InputData.java
@@ -44,7 +44,7 @@ public class InputData implements Serializable{
 	private Double  crackLength; // 裂纹形状比
 	private Double  relativeLength; // 裂纹相对深度
 	
-	private Integer pressure;  // 内压
+	private Double pressure;  // 内压
 	
 	private Integer yield; // 屈服强度
 	
@@ -139,14 +139,14 @@ public class InputData implements Serializable{
 	public void setRelativeLength(Double relativeLength) {
 		this.relativeLength = relativeLength;
 	}
-	public Integer getPressure() {
+
+
+	public Double getPressure() {
 		return pressure;
 	}
-
-	public void setPressure(Integer pressure) {
+	public void setPressure(Double pressure) {
 		this.pressure = pressure;
 	}
-
 	public User getUser() {
 		return user;
 	}
